Add block-tool assertion and tool switching in hotbar
